rename helpers file to mapFields
