+++ src/index.minimal.esm.js
@@ -4,7 +4,7 @@ import mixin from './mixin';
 import install from './install';
 import use from './use';
 import ErrorBag from './core/errorBag';
-import mapFields from './core/helpers';
+import mapFields from './core/mapFields';
 
 const version = '__VERSION__';
